Changes to support Swift 1.2
Many of these changes are related to the use of @autoclosure.

Swift 1.2 introduced the @noescape attribute which can be used with closures to indicate that the closure will not outlive the lifetime of the call. A @noescape closure will not be assigned and stored in another variable or be called later. The compiler ensures it is only ever called or passed as a parameter to another function expecting a @noescape closure. When using @noescape closures, the caller doesn't need to worry about how variables are captured as the compiler guarantees that the closure will not outlive the call.

By default @autoclosures are now @noescape closures. Apple have said (in the Swift 1.2 release notes) "This intentionally limits the power of @autoclosure to control-flow and lazy evaluation use cases". I'm not familiar with why @autoclosures are used in Haneke, but this change in Swift may mean their use needs to be reconsidered. In the meantime, Swift allows you to revert to the previous behaviour by annotating the closure with @autoclosure(escaping) to make it clear that the @autoclosure in question may outlive the life of the call.